Show output folder for exported files
Added display of the output folder (because it's not at all obvious where it's going to end up, and if you close the window that auto-opens at end of process - it's gone! Also: Windows10 has a habit of not-repopening Explorer windows that were already open, making it hard to find the file with current design).

Also changes the success message to give the actual filename/location (previous version only gives a relative filepath, which can't be copy/pasted, can't be opened directly, etc)